+++ packages/af-virtual-scroll/src/hooks/useVirtual/index.js
@@ -16,7 +16,7 @@ const useVirtual = ({
         const model = new ( fixed ? FixedHeightsModel : VariableHeightsModel );
 
         if( typeof window === "undefined" ){
-            model.widgetSize = ssrWidgetSize;
+            model._widgetSize = ssrWidgetSize;
         }
         
         model._setParams( estimatedItemSize, overscanCount, itemCount, horizontal );