+++ packages/website/components/examples/list/VariableSizeList/index.js
@@ -10,7 +10,10 @@ const DEFAULT_ROW_COUNT = 2000;
 
 const VariableSizeList = () => {
 
-    const [ dynamicListRowHeights ] = useState(() => times( DEFAULT_ROW_COUNT, () => r( 50, 100 ) ));
+    const [ dynamicListRowHeights ] = useState(() => times(
+        DEFAULT_ROW_COUNT,
+        () => r( 50, 100 )
+    ));
 
     const model = useVirtual({
         itemCount: DEFAULT_ROW_COUNT
@@ -19,10 +22,14 @@ const VariableSizeList = () => {
     return (
         <VerticalList model={model}>
             {i => (
-                <div key={i} className="text-center border-t border-zinc-800" style={{
-                    lineHeight: `${dynamicListRowHeights[i]}px`,
-                    background: `hsl(${i*11%360},60%,60%)`
-                }}>
+                <div
+                    key={i}
+                    className="text-center border-t border-zinc-800"
+                    style={{
+                        lineHeight: `${dynamicListRowHeights[i]}px`,
+                        background: `hsl(${i*11%360},60%,60%)`
+                    }}
+                >
                     row {i}:&nbsp;{dynamicListRowHeights[i]}px
                 </div>
             )}