+++ packages/af-react-table/CHANGELOG.md
@@ -1,5 +1,14 @@
 # af-react-table
 
+## 4.0.2
+
+### Patch Changes
+
+-   added 'horizontal' prop to useVirtual
+
+-   Updated dependencies []:
+    -   af-virtual-scroll@2.10.2
+
 ## 4.0.1
 
 ### Patch Changes