+++ packages/website/CHANGELOG.md
@@ -1,5 +1,15 @@
 # website
 
+## 1.1.2
+
+### Patch Changes
+
+-   added 'horizontal' prop to useVirtual
+
+-   Updated dependencies []:
+    -   af-virtual-scroll@2.10.2
+    -   af-react-table@4.0.2
+
 ## 1.1.1
 
 ### Patch Changes